test without new_transactions
