+++ chain/chain/Cargo.toml
@@ -81,6 +81,9 @@ no_cache = ["near-store/no_cache"]
 protocol_feature_reject_blocks_with_outdated_protocol_version = [
   "near-primitives/protocol_feature_reject_blocks_with_outdated_protocol_version",
 ]
+protocol_feature_relaxed_chunk_validation = [
+  "near-primitives/protocol_feature_relaxed_chunk_validation",
+]
 
 nightly = [
   "near-async/nightly",
@@ -98,6 +101,7 @@ nightly = [
   "nightly_protocol",
   "node-runtime/nightly",
   "protocol_feature_reject_blocks_with_outdated_protocol_version",
+  "protocol_feature_relaxed_chunk_validation",
 ]
 nightly_protocol = [
   "near-async/nightly_protocol",
