+++ core/primitives-core/Cargo.toml
@@ -38,6 +38,7 @@ protocol_feature_fix_staking_threshold = []
 protocol_feature_fix_contract_loading_cost = []
 protocol_feature_reject_blocks_with_outdated_protocol_version = []
 protocol_feature_nonrefundable_transfer_nep491 = []
+protocol_feature_relaxed_chunk_validation = []
 
 nightly = [
   "nightly_protocol",
@@ -45,6 +46,7 @@ nightly = [
   "protocol_feature_fix_staking_threshold",
   "protocol_feature_nonrefundable_transfer_nep491",
   "protocol_feature_reject_blocks_with_outdated_protocol_version",
+  "protocol_feature_relaxed_chunk_validation",
 ]
 
 nightly_protocol = [
