+++ integration-tests/src/test_loop/tests/mod.rs
@@ -1,19 +1,19 @@
-pub mod bandwidth_scheduler_protocol_upgrade;
+mod bandwidth_scheduler_protocol_upgrade;
 mod chunk_validator_kickout;
-pub mod congestion_control;
-pub mod congestion_control_genesis_bootstrap;
-pub mod contract_distribution_cross_shard;
-pub mod contract_distribution_simple;
+mod congestion_control;
+mod congestion_control_genesis_bootstrap;
+mod contract_distribution_cross_shard;
+mod contract_distribution_simple;
 mod create_delete_account;
-pub mod epoch_sync;
-pub mod fix_min_stake_ratio;
-pub mod in_memory_tries;
-pub mod max_receipt_size;
-pub mod multinode_stateless_validators;
-pub mod multinode_test_loop_example;
-pub mod protocol_upgrade;
+mod epoch_sync;
+mod fix_min_stake_ratio;
+mod in_memory_tries;
+mod max_receipt_size;
+mod multinode_stateless_validators;
+mod multinode_test_loop_example;
+mod protocol_upgrade;
 mod resharding_v3;
-pub mod simple_test_loop_example;
-pub mod state_sync;
-pub mod syncing;
-pub mod view_requests_to_archival_node;
+mod simple_test_loop_example;
+mod state_sync;
+mod syncing;
+mod view_requests_to_archival_node;
