+++ integration-tests/src/tests/client/mod.rs
@@ -5,6 +5,7 @@ mod chunks_management;
 mod cold_storage;
 mod features;
 mod flat_storage;
+mod invalid_txs;
 mod process_blocks;
 mod resharding_v2;
 mod runtimes;
